Explicitly pass, document, and limit the variables accepted by slim_header.html
We'll be building this as the replacement for header.html eventually.

So, we should properly document each variable and make sure the template
is used as intended by not implicitly passing all variables.

Unfortunately, this isn't possible yet for pages and snippets due to the
side panels relying on some variables from the view.