Move filtering logic from views.generic.models.IndexView to views.generic.base.BaseListingView
BaseListingView now supports passing a django-filters filterset class as `self.filterset_class`, provides properties `filters` and `is_filtering`, and provides a `filter_queryset` method. If a filterset class is specified, this will be used to filter the queryset within `get_queryset`; `filters` and `is_filtering` will be added to the template context; and the `media` context variable will include the filter form's media.