Move locale selector in generic IndexView to a filter
Ensure the locale filter is always included by overriding
WagtailFilterSet.get_filters(), so if developers use a custom
filterset_class, the locale filter will automatically be added.

Adding an initial default value for a filter is not recommended as it
prevents users from skipping the filter:

https://django-filter.readthedocs.io/en/stable/guide/tips.html#using-initial-values-as-defaults

For this reason, add the option to filter by "All" locales instead,
which is applied when the locale filter is empty.

This means that the default state of the IndexView (before a locale is
specifically selected) is now to show objects from all locales instead
of only those of the default locale.

All this is only done if breadcrumbs are used, which means that the slim
header (along with its search+filter components) is used.

If breadcrumbs are not used i.e. the legacy header is still in use,
keep the old logic with the legacy locale selector for now.