Remove legacy locale selector fallback in generic IndexView
This means that for index views that don't use the breadcrumbs,
locale selection is now done using the filters on the right side.

Remove the locale filtering from get_queryset(), as it's been
handled by filter_queryset() using django-filter.

Also remove the get_translations() method as it was only used to
generate links for the legacy locale selector.