Move handling of construct_explorer_page_queryset hook into search_queryset
This makes get_queryset more-or-less generic. It's a bit of a hack to put it here, but it's reasonable to say that this should happen immediately before the search logic (so that the queryset passed to the hook is as close as possible to the final result, while still actually being a queryset and not a SearchResults object).