Use w-swap for removing active filters
This allows the filters to be removed via AJAX. However, the filter form
won't reflect the updates automatically, as it's not part of the results
response. This should be fixed separately, once we have the filter popup
component in place.

In addition, also simulate the future intended behaviour of opening the
active filter when clicking on the filter (not the remove button) by
opening the filters dialog instead.