Fix nice-padding of before_results block in images listing
Without this fix, the "other searches bar" will have the nice-padding
applied twice, making it appear too close to the centre.

Unfortunately, we have to use a new nice-padding wrapper instead of just
applying the class to the <ul>, because we apply the "unlist" mixin to
ul.listing which includes setting the left padding to 0 (and it beats
the specificity of a single .nice-padding class).