Document MIT licensing terms for handsontable 6.2.2
Ref: #11497
People regularly raise concerns about handsontable as a commercial product being incompatible with Wagtail's licence. Clarify that we are using the last version available under the MIT licence, and link to the relevant git tag to allow people to verify this (since the handsontable website no longer mentions the open-source version, and is probably a vastly different product at this point).