Enforce page_breadcrumbs to be expanded when there is only one item
Rather than special-casing whether the page is root or not, just check
if there is only one item. If that's the case, then the user is
exploring the "explorable root instance", which may be the root Page
instance or the closest common ancestor where the user has permission.

In this case, we don't want to show the breadcrumbs toggle/expand button
(it's always expanded).

This also removes the incorrectly extraneous right arrow when exploring
the root instance.