Add initial Stimulus FormsetController `w-formset`
Replaces the core functionality in `client/src/components/ExpandingFormset/index.js` and the Delete callbacks in other usage of the `window.buildExpandingFormset` function.

Provides limitations on add/delete once the min/max have been reached but does not set up any disabling of these buttons as this is currently in the remit of `InlinePanel`.

Includes a full test suite for error handling, updating inputs and nested usage which leverages Stimulus' scope management.

Does not rely on any id attribute structure or wrangling.

Relates to #7915