Move modal-workflow.js script to admin_base.html (#11620)
The modal-workflow.js script is used across a wide range of views in the admin, it's simpler to include it globally instead of having to consider all places it may be used.

Also unblocks use cases that are outside of the core admin such as custom documents/user/image models that may want to leverage this functionality.

Includes a fix for the documented support for all Panels in Wagtail 6.0's release notes, even though DocumentChooser and others would not have worked.

Cherry pick of single commit 266faf63c2967f01c7be01afb0fdaf92bc77c5c4 from #11620

Backport fix for #11739