Move modal-workflow.js script to admin_base.html
The modal-workflow.js script is used across a wide range of views in the admin, it's simpler to include it globally instead of having to consider all places it may be used.

Also unblocks use cases that are outside of the core admin such as custom documents/user/image models that may want to leverage this functionality.