Move 'sort by' select element in images listing to the slim header
This uses a custom header template that extends the slim_header.html

We could do this without the custom header template, but this would mean
the slim_header.html will accept yet another variable (extra_form_fields),
which we are trying to avoid to prevent it from accepting so many variables
like header.html.

This is especially since this "extra fields" block is only needed for the
images listing (at least for the foreseeable future).