Show full label for non-page listings HeaderButton
In the new designs, listings other than the page listings should show
the complete label for the main header button.

The page listing only shows the icon because much of the space is taken
by the side panel toggles.

The page listing's header button hasn't been refactored to use the
HeaderButton class, hence there is no changes to the page listings code
in this commit.