Use params in the element's URL as the default for LinkController.reflectKeysValue
With this change, we no longer need to manually specify the
data-w-link-reflect-keys-value attribute from the server. As long as the
get_add_url method returns the URL with the desired parameters to be
reflected (even with empty values), those parameters will be used by
LinkController so that it reflects any updates to those params when needed.

Any elements that want to reflect all params need to explicitly set the
__all__ value.