Use cached_property for RevisionMixin.revisions and WorkflowMixin.workflow_states
For models that have multi-table inheritance, accessing a
GenericRelation will not always give you the correct results. This can
occur if the content type used by the GenericRelation does not match the
instance at hand, e.g. accessing a GenericRelation that uses
base_content_type from an instance of the specific class, or vice versa.

This is a known quirk when using GenericRelation together with MTI:
https://code.djangoproject.com/ticket/31269

To work around this, we need to have a method or property that resolves
to the correct GenericRelation (or the related model's QuerySet) to use,
depending on whether the current instance is specific or not. This
allows us to use and publicly document RevisionMixin.revisions as
something that always gives you the correct queryset of Revisions for
the object, without having to deal with the MTI quirks.

This is why the RevisionMixin has a `revisions` property that queries
the Revision model directly, so that accessing MyModelInstance.revisions
always gives you the correct results, no matter how the GenericRelation
is set up for custom models.

For pages, we have access to the model implementation, so the
Page.revisions property is implemented using a GenericRelation that is
always accessed using a specific version of the instance.

For models that don't use MTI, we advise developers to just override the
revisions property with a GenericRelation definition. However, as of
https://github.com/django/django/commit/faeb92ea13f0c1b2cc83f45b512f2c41cfb4f02d,
the existence of a @property and a related field accessor of the same
name triggers a system check error.

It is unknown whether the system check error is the desired result or
not, as the error wasn't triggered before Django made the optimisation.
The approach that we documented about shadowing the mixins' default
properties with a real GenericRelation clearly worked before though, so
it is probably okay.

To work around the new Django behaviour, we can use cached_property
instead, which bypasses the check, as Django only checks for the plain
Python `property`.

Alternatively, if the system check error is indeed the desired result,
we can suggest developers to **never** override the `revisions` and
`workflow_states` with `GenericRelation`s directly, and instead define
the `GenericRelation`s with other names, and override the property as
a new property that returns the appropriate `GenericRelation` (with
any additional logic as necessary, e.g. to handle MTI).