Fix dummy request SERVER_NAME when running tests
Django's test runner appends 'testserver' to your ALLOWED_HOSTS
setting. If your project has `ALLOWED_HOSTS=['*']` (which is totally
legit for Google App Engine standard) this means the setting value is
`ALLOWED_HOSTS=['*', 'testserver']` and Wagtail's dummy request helper
was setting the request SERVER_NAME to '*'.

But '*' is not a valid host name, causing a DisallowedHost exception.

This change sets the SERVER_NAME to 'example.com' in that case.