Fix timezone handling in timesince_last_update
Ref: https://github.com/wagtail/wagtail/pull/11815/files#r1556095493

test_group_edit was failing when run after 3pm UTC, due to the output being rendered as "Just now" instead of an explicit time. This was due to a previously unsurfaced logic error in the timesince_last_update tag: if the timestamp was passed as a timezone-aware datetime in UTC, it was failing to translate this to local time and thus it compared as a different date, even if it was only from a moment ago. Consequently, it was following the "X days ago" path that was only supposed to be used for timestamps before today.

With this fixed, the special-case for "Just now" can be removed since it won't come into play for timestamps of yesterday and earlier. (Well, technically it could in the first minute after midnight, but that really doesn't warrant a special case...)