Tweak styles for w-header-button to better match the designs
Ideally, the hover state of the icon should be implemented by using the
regular variant and switching to the solid variant on hover, instead of
manually setting the color and background color. And ideally, the add
button should use the FA circle-plus icon instead of the plus icon.
However, this is how the StreamField add button is styled, which is the
closest existing implementation of the styles.

Remove the --icon-only variant added in
88fad9a07143fd3872992fe6f7e0e2f5841fc2ef as it is now redundant. The
styles now work for both icon-only and labelled buttons.

Adjust the font-weight and spacing per feedback from Ben.