Render listing buttons as components
Instead of manually rendering them via the parent dropdown component and
a custom template, delegate each button's rendering to itself by making
use of the default Button template added in
2274a44572b6b3d8a19bd4a1d4c893ab2b0d51ca and the {% component %} tag.

This allows the buttons inside the dropdown to have their own markup,
e.g. to use a <button> element instead of <a>.

Suggest people to move away from PageListingButton to the new PageButton
or the existing plain Button class for dropdown button hooks. In the
meantime, remove the `button button-small button-secondary` CSS classes
from the button instance if it's used in a dropdown.

This commit mainly focuses on the page buttons, and adding a test to the
users index view where a custom button component is used to ensure the
buttons are rendered as components.

Usages of the default ListingButton in the generic IndexView, as well as
SnippetListingButton and UserListingButton, are broken with this commit.
This will be fixed in the next commit, where we'll do the same
modification to the classname and add the formal deprecations.