Render listing buttons as components
Instead of manually rendering them via the parent dropdown component and
a custom template, delegate each button's rendering to itself by making
use of the default Button template added in
2274a44572b6b3d8a19bd4a1d4c893ab2b0d51ca and the {% component %} tag.

This allows the buttons inside the dropdown to have their own markup,
e.g. to use a <button> element instead of <a>.

Remove the `button button-small button-secondary` CSS classes from the
base ListingButton class. We'll assume instances of this class is
rendered inside a dropdown, and hence do not need the extra CSS. If
people still do need it, advise them to add it to the classname
themselves.