Partially revert listing styles changes in b8dd7f4, 08ee15a, and bb12877
In bb12877, the change of .title-wrapper from inline to inline-flex
caused an issue with Safari, which b8dd7f4 tried to fix.

Then, further cleanup in 08ee15a caused an issue to the title cell.
While we intended the styles to vertically align the title cells, this
wasn't a good idea for a few reasons:

- The styles get applied to both td and th elements
- Even if we limit this to td.title, setting display: flex directly on
  the td element is bad, as it would cause the table layout to break in
  some cases (e.g. if two td.title elements are adjacent). This issue
  technically also existed in the old `--inline-actions` styles before
  08ee15a, but it was pretty isolated as only `HistoryView` used it.

bb12877 was the commit that started it all. The styling changes in there
was to ensure the user avatar and the user name is vertically aligned.
However, since the changes in that commit caused a chain reaction to
this point, I'll try a different approach instead. Using inline-flex
instead of flex for the user avatar + name wrapper seems to do the
trick.