Remove flex styles from .listing .title
While we intend these styles to vertically align the title cells, this
wasn't a good idea for a few reasons:

- The styles get applied to both td and th elements
- Even if we limit this to td.title, setting display: flex directly on
  the td element is bad, as it would cause the table layout to break in
  some cases (e.g. if two td.title elements are adjacent)

Remove these styles for now. This doesn't cause too much issue, since
most things that need to be vertically centered are usually in a
.title-wrapper element, which we have fixed in
b8dd7f484f428e8027dc6993b113084dfa539986. The only issue is that now the
"..." actions button may not be vertically centered when the title is
long. That will need to be fixed by introducing another wrapper that
wraps the .title-wrapper and the .actions elements together.