Create ModelPermissionTester based on PagePermissionTester
Use the permission policies user_has_permission_for_instance under the
hood as much as possible, making sure that we implement the basic checks
in there whenever possible.

This is different to PagePermissionTester that only uses the policy to
query the GroupPagePermission objects, but still does all of the actual
permission checks by itself. This was because it precedes the policy,
and we wanted to avoid making too much breakage.