Change site history's commenting action filter to be a tri-state filter
This allows users to also show commenting actions *only*. In addition,
this also fixes an issue where the active filter mechanism thinks that
the hide_commenting_action filter is active with the value False when
the filter is left unchecked.

The test for the inactive filter state is already covered by
test_unfiltered.

Similar to 2173a42345081f4b94f5844fafa8e3823725407f