Use PermissionCheckedMixin in ReportView
Also fix an issue where the workflow tasks report is accessible despite
the user does not have the necessary page permissions, because the fix
in 13048eab793d80f6bc7a32aa14a1c3511f494fa9 is missing the view
dispatch() override for the workflow task report view.

This wasn't a security risk as the queryset is filtered down to the data
where the user has the relevant permissions for.