Rename pulse-warning animation to ping-warning and extract to its own file
Tailwind has a similarly-named 'pulse' animation that behaves differently:
it makes the element fade in and out.

Meanwhile, it has a 'ping' animation that behaves similarly to our 'pulse'
animation: https://tailwindcss.com/docs/animation

Follow the Tailwind names to avoid confusion. In the future, we might
want to implement a similar 'pulse' animation. Or, perhaps we can reuse
Tailwind's animations as-is.