Move sprite hashing out of module import time
This speeds up application startup. The hash is now a query param, injected in the template. As this param is only needed for cache invalidation, it's optional. A helper method is provided to generate the URL, along with a template tag.

This also migrates to an `lru_cache` over a global variable for simplicity.

Fixes #11680