Use a snippet model without RevisionMixin for testing EditingSession
We want to make sure the feature works (without the new version
detection) when using a snippet model that does not use RevisionMixin.

We don't have any tests that ensures the new version detection works
with a RevisionMixin-enabled snippet. We can add those later, but the
code itself doesn't make a special-case between snippet vs pages, so
it's probably okay to skip them.