Fix flaky SwapController tests
And change the description for one of the describe blocks, which may
have been a copy-paste leftover as there's an existing describe block
with the exact same sentence. Change this to accurately describe that
we are testing without the form's input values (as we're using
replace() directly instead of submit()).