Do not suggest using the custom user model's app config as the custom wagtail.users app config
Doing so would mean the "models" module of the app that contains the
custom user model will be used in favour of the models module of the
wagtail.users app, which contains the UserProfile model. As a result,
the UserProfile model becomes mistakenly picked up by the wagtailcore
app, creating a bogus migration when makemigrations is run.

This partially reverts the docs to the version before
449a48d7f9816fa17eda2ca258de41ed13df8e39, in particular the paragraphs
about creating the custom AppConfig subclass for wagtail.users.