Move publish, lock, and unlock permissions to 'Custom permissions' column
These permissions only show up for models that have the corresponding
mixins enabled, which are likely to be the minority – which results in
the other rows having an empty cell for the column. This takes up space
in an already crowded table.

Move them to the "Custom permissions" column so we can reclaim the space
for other, more important columns.