Use approved_schedule property for detecting ScheduledForPublishLock
We only need to know about the existence of a scheduled revision,
and not the revision itself.

This method is used by 'PagePermissionTester.can_unpublish', which
in turn is used by the 'Unpublish' button in the page listing 'more'
button. As a result, this method is run on every single item in the
listing.

We already have annotate_approved_schedule to optimise the display
of the 'scheduled' status tag (for pages). Use the approved_schedule
property so we can reuse that annotation instead of performing an
N+1 query.