Fix RevisionQuerySet.for_instance() when used with a non-specific instance
If the method was used with a base Page instance, it wouldn't return any
revisions because we would be filtering on the content_type FK using the
current model (the base Page model) instead of the specific model.

Filter on base_content_type and make use of
RevisionMixin.get_base_content_type() instead, which will
resolve to the base Page model (and the correct the most basic
non-abstract model for non-Page models with MTI).

Use the old logic if the instance's model does not use RevisionMixin for
some reason.

This logic is similar to WorkflowStateQuerySet.for_instance() and
TaskQuerySet.for_instance().