Add tests for rendering workflow history detail view
and update the docstring of the revision query to reflect the actual
behaviour. The query does:

- a filter on TaskState to get the ones that are associated with the
  current workflow state, then gets the IDs of the revisions associated
  with those task states, and
- use it as a subquery to get the revisions with those IDs

TaskState objects are only created when a task's state changes – this
means we don't create new TaskState objects when you save the
page/snippet without performing a workflow action.

This means the query will only result in revisions that were made as
part of a workflow action, not just any edits.